requirements.txt
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
Flask==2.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
neo4j==5.3.0
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==0.21.0
tomli==2.0.2
typing_extensions==4.12.2
Werkzeug==2.0.3
WTForms==3.2.1
neo4j==5.27.0
